separate audit log from system log and refactor to share download code
